requirements/core.txt
# From the cis_production lambda layer.
aniso8601==7.0.0
async_generator==1.10
attrs==20.2.0
auth0_python==3.13.0
aws_xray_sdk==2.6.0
boto3==1.34.162
botocore==1.34.162
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
configobj==5.0.9
cryptography==45.0.5
ecdsa==0.14.1
everett==1.0.2
Faker==4.1.3
Flask_Cors==3.0.9
Flask_RESTful==0.3.8
Flask==1.1.2
Flask-GraphQL==2.0.1
graphene==2.1.8
graphql_core==2.3.2
graphql_relay==2.0.1
graphql-server-core==1.2.0
iam_profile_faker==0.3.1
idna==2.10
importlib_metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
orjson==3.11.0
jsonpickle==1.4.1
jsonschema==3.2.0
MarkupSafe==1.1.1
outcome==1.0.1
#psycopg2==2.9.10
psycopg2-binary==2.9.10
pyaml==20.4.0
pyasn1==0.4.8
pycparser==2.20
python_dateutil==2.8.1
python_jose==3.2.0
python-json-logger==3.3.0
pytz==2020.1
requests==2.32.4
rsa==4.6
Rx==1.6.1
s3transfer==0.10.4
serverless_wsgi==1.7.5
setuptools==50.3.0
six==1.15.0
sniffio==1.1.0
sortedcontainers==2.2.2
SQLAlchemy_Utils==0.41.2
SQLAlchemy==1.3.19
text_unidecode==1.3
tinydb==4.1.1
trio==0.17.0
typing_extensions==4.14.0
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.2.0
